Use useRenderElement
