+++ packages/react/src/collapsible/panel/CollapsiblePanel.tsx
@@ -1,8 +1,8 @@
 'use client';
 import * as React from 'react';
 import { BaseUIComponentProps } from '../../utils/types';
-import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import { useModernLayoutEffect } from '../../utils/useModernLayoutEffect';
+import { useRenderElement } from '../../utils/useRenderElement';
 import { warn } from '../../utils/warn';
 import { useCollapsibleRootContext } from '../root/CollapsibleRootContext';
 import type { CollapsibleRoot } from '../root/CollapsibleRoot';
@@ -18,7 +18,7 @@ import { usePanelResize } from '../../utils/usePanelResize';
  * Documentation: [Base UI Collapsible](https://base-ui.com/react/components/collapsible)
  */
 const CollapsiblePanel = React.forwardRef(function CollapsiblePanel(
-  props: CollapsiblePanel.Props,
+  componentProps: CollapsiblePanel.Props,
   forwardedRef: React.ForwardedRef<HTMLDivElement>,
 ) {
   const {
@@ -27,8 +27,8 @@ const CollapsiblePanel = React.forwardRef(function CollapsiblePanel(
     id: idProp,
     keepMounted: keepMountedProp,
     render,
-    ...otherProps
-  } = props;
+    ...elementProps
+  } = componentProps;
 
   if (process.env.NODE_ENV !== 'production') {
     // eslint-disable-next-line react-hooks/rules-of-hooks
@@ -75,7 +75,7 @@ const CollapsiblePanel = React.forwardRef(function CollapsiblePanel(
     setKeepMounted(keepMounted);
   }, [setKeepMounted, keepMounted]);
 
-  const { getRootProps } = useCollapsiblePanel({
+  const { props } = useCollapsiblePanel({
     abortControllerRef,
     animationTypeRef,
     externalRef: forwardedRef,
@@ -100,24 +100,28 @@ const CollapsiblePanel = React.forwardRef(function CollapsiblePanel(
 
   usePanelResize(panelRef, setDimensions, open);
 
-  const { renderElement } = useComponentRenderer({
-    propGetter: getRootProps,
-    render: render ?? 'div',
+  const renderElement = useRenderElement('div', componentProps, {
     state,
-    className,
     ref: [forwardedRef, panelRef],
-    extraProps: {
-      ...otherProps,
-      style: {
-        ...otherProps.style,
-        [CollapsiblePanelCssVars.collapsiblePanelHeight]: height ? `${height}px` : undefined,
-        [CollapsiblePanelCssVars.collapsiblePanelWidth]: width ? `${width}px` : undefined,
+    props: [
+      props,
+      {
+        style: {
+          [CollapsiblePanelCssVars.collapsiblePanelHeight as string]: height
+            ? `${height}px`
+            : undefined,
+          [CollapsiblePanelCssVars.collapsiblePanelWidth as string]: width
+            ? `${width}px`
+            : undefined,
+        },
       },
-    },
+      elementProps,
+    ],
     customStyleHookMapping: collapsibleStyleHookMapping,
   });
 
   const shouldRender = keepMounted || hiddenUntilFound || (!keepMounted && mounted);
+
   if (!shouldRender) {
     return null;
   }
